Move internal field to its own file
